refactor(generator): rename InsertColumnInfo -> WriteColumnInfo
Insert and patch now share the same per-column metadata type, so the
Insert prefix on the column model became misleading. Renamed it to
WriteColumnInfo and pulled it out of Models/InsertInfo.cs into its own
Models/WriteColumnInfo.cs.

Consumers updated: InsertInfo.Columns, PatchInfo.Columns, and
TerminalEmitHelpers.GetInsertColumnBinding. The type is internal —
no external API change.

Tests: 3,503 passing (no behavior change).
