+++ src/Quarry.Generator/Models/InsertInfo.cs
@@ -11,7 +11,7 @@ internal sealed class InsertInfo : IEquatable<InsertInfo>
     /// <summary>
     /// Gets the columns to insert (excluding Identity and Computed columns).
     /// </summary>
-    public IReadOnlyList<InsertColumnInfo> Columns { get; }
+    public IReadOnlyList<WriteColumnInfo> Columns { get; }
 
     /// <summary>
     /// Gets the identity column name (for RETURNING clause), or null if none.
@@ -30,7 +30,7 @@ internal sealed class InsertInfo : IEquatable<InsertInfo>
     public string? QuotedIdentityColumnName { get; }
 
     public InsertInfo(
-        IReadOnlyList<InsertColumnInfo> columns,
+        IReadOnlyList<WriteColumnInfo> columns,
         string? identityColumnName,
         string? identityPropertyName,
         string? quotedIdentityColumnName = null)
@@ -46,7 +46,7 @@ internal sealed class InsertInfo : IEquatable<InsertInfo>
     /// </summary>
     public static InsertInfo FromEntityInfo(EntityInfo entity, SqlDialect dialect, System.Collections.Generic.HashSet<string>? initializedPropertyNames = null)
     {
-        var columns = new List<InsertColumnInfo>();
+        var columns = new List<WriteColumnInfo>();
         string? identityColumnName = null;
         string? identityPropertyName = null;
 
@@ -69,7 +69,7 @@ internal sealed class InsertInfo : IEquatable<InsertInfo>
             if (initializedPropertyNames != null && !initializedPropertyNames.Contains(column.PropertyName))
                 continue;
 
-            columns.Add(new InsertColumnInfo(
+            columns.Add(new WriteColumnInfo(
                 propertyName: column.PropertyName,
                 columnName: column.ColumnName,
                 quotedColumnName: FormatColumnName(column.ColumnName, dialect),
@@ -121,148 +121,5 @@ internal sealed class InsertInfo : IEquatable<InsertInfo>
     public override int GetHashCode()
     {
         return HashCode.Combine(IdentityColumnName, IdentityPropertyName, Columns.Count);
-    }
-}
-
-/// <summary>
-/// Information about a column in an insert operation.
-/// </summary>
-internal sealed class InsertColumnInfo : IEquatable<InsertColumnInfo>
-{
-    /// <summary>
-    /// Gets the property name in the entity class.
-    /// </summary>
-    public string PropertyName { get; }
-
-    /// <summary>
-    /// Gets the database column name.
-    /// </summary>
-    public string ColumnName { get; }
-
-    /// <summary>
-    /// Gets the quoted column name for SQL generation.
-    /// </summary>
-    public string QuotedColumnName { get; }
-
-    /// <summary>
-    /// Gets the CLR type name.
-    /// </summary>
-    public string ClrType { get; }
-
-    /// <summary>
-    /// Gets the fully qualified CLR type name.
-    /// </summary>
-    public string FullClrType { get; }
-
-    /// <summary>
-    /// Gets whether the column is nullable.
-    /// </summary>
-    public bool IsNullable { get; }
-
-    /// <summary>
-    /// Gets whether the CLR type is a value type.
-    /// </summary>
-    public bool IsValueType { get; }
-
-    /// <summary>
-    /// Gets whether this column is a foreign key (Ref&lt;TEntity, TKey&gt;).
-    /// When true, generated code must extract .Id before passing to ADO.NET.
-    /// </summary>
-    public bool IsForeignKey { get; }
-
-    /// <summary>
-    /// Gets the referenced entity type name for FK columns (e.g., "User").
-    /// Null for non-FK columns.
-    /// </summary>
-    public string? ForeignKeyEntityName { get; }
-
-    /// <summary>
-    /// Gets the fully qualified custom TypeMapping class name, if this column uses one.
-    /// When set, insert code should wrap the value with mapper.ToDb(value).
-    /// </summary>
-    public string? CustomTypeMappingClass { get; }
-
-    /// <summary>
-    /// Gets whether this column contains sensitive data.
-    /// When true, parameter values are redacted in log output.
-    /// </summary>
-    public bool IsSensitive { get; }
-
-    /// <summary>
-    /// Gets whether the CLR type is an enum.
-    /// When true, carrier insert code must cast to the underlying integer type.
-    /// </summary>
-    public bool IsEnum { get; }
-
-    /// <summary>
-    /// Gets whether the CLR type is a boolean.
-    /// When true, carrier insert code must convert to 0/1 and set DbType.Int32
-    /// for providers (e.g. SQLite) that reject boxed booleans.
-    /// </summary>
-    public bool IsBoolean { get; }
-
-    /// <summary>
-    /// Gets the underlying type name for enum columns (e.g., "int", "byte").
-    /// Null for non-enum columns. Used for the cast in generated insert code.
-    /// </summary>
-    public string? EnumUnderlyingType { get; }
-
-    public InsertColumnInfo(
-        string propertyName,
-        string columnName,
-        string quotedColumnName,
-        string clrType,
-        string fullClrType,
-        bool isNullable,
-        bool isValueType,
-        bool isForeignKey = false,
-        string? foreignKeyEntityName = null,
-        string? customTypeMappingClass = null,
-        bool isSensitive = false,
-        bool isEnum = false,
-        bool isBoolean = false,
-        string? enumUnderlyingType = null)
-    {
-        PropertyName = propertyName;
-        ColumnName = columnName;
-        QuotedColumnName = quotedColumnName;
-        ClrType = clrType;
-        FullClrType = fullClrType;
-        IsNullable = isNullable;
-        IsValueType = isValueType;
-        IsForeignKey = isForeignKey;
-        ForeignKeyEntityName = foreignKeyEntityName;
-        CustomTypeMappingClass = customTypeMappingClass;
-        IsSensitive = isSensitive;
-        IsEnum = isEnum;
-        IsBoolean = isBoolean;
-        EnumUnderlyingType = enumUnderlyingType;
-    }
-
-    public bool Equals(InsertColumnInfo? other)
-    {
-        if (other is null) return false;
-        if (ReferenceEquals(this, other)) return true;
-        return PropertyName == other.PropertyName
-            && ColumnName == other.ColumnName
-            && QuotedColumnName == other.QuotedColumnName
-            && ClrType == other.ClrType
-            && FullClrType == other.FullClrType
-            && IsNullable == other.IsNullable
-            && IsValueType == other.IsValueType
-            && IsForeignKey == other.IsForeignKey
-            && ForeignKeyEntityName == other.ForeignKeyEntityName
-            && CustomTypeMappingClass == other.CustomTypeMappingClass
-            && IsSensitive == other.IsSensitive
-            && IsEnum == other.IsEnum
-            && IsBoolean == other.IsBoolean
-            && EnumUnderlyingType == other.EnumUnderlyingType;
-    }
-
-    public override bool Equals(object? obj) => Equals(obj as InsertColumnInfo);
-
-    public override int GetHashCode()
-    {
-        return HashCode.Combine(PropertyName, ColumnName, ClrType, IsNullable);
     }
 }
